inference/gradio_demo/requirements.txt
absl-py==0.15.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.2.1
anyio==3.6.2
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
av==9.2.0
bcrypt==4.0.1
bleach==1.5.0
blessings==1.7
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
conda-pack==0.6.0
cryptography==38.0.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
decorator==4.4.2
docker-pycreds==0.4.0
dominate==2.6.0
einops==0.4.1
enum34==1.1.10
face-alignment==1.3.4
fastapi==0.88.0
ffmpy==0.3.0
flatbuffers==1.12
fonttools==4.27.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gast==0.3.3
gitdb==4.0.9
GitPython==3.1.20
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==0.6.0
gradio==3.12.0
graphviz==0.8.4
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
html5lib==0.9999999
httpcore==0.15.0
httpx==0.23.1
idna==3.3
idna-ssl==1.1.0
imageio==2.15.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
imutils==0.5.4
inplace-abn==1.1.1.dev6+gd2728c8.d20220908
Jinja2==3.1.2
joblib==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kornia==0.5.8
linkify-it-py==1.0.3
llvmlite==0.36.0
lmdb==1.3.0
lpips==0.1.4
Markdown==3.3.6
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==5.2.0
mxnet-cu110==1.9.1
networkx==2.5.1
neuralgym==0.0.1
ninja==1.10.2.3
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
nvdiffrast==0.2.7
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
onnxruntime==1.10.0
opencv-python==4.5.5.62
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
orjson==3.8.3
packaging==21.3
pandas==1.1.5
paramiko==2.12.0
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pyDeprecate==0.3.1
pydub==0.25.1
PyNaCl==1.5.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-multipart==0.0.5
pytorch-lightning==1.4.5
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
sentry-sdk==1.5.6
setproctitle==1.2.2
shortuuid==1.0.8
six==1.15.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.5.4
torch==1.7.1+cu110
torchmetrics==0.7.2
torchvision==0.8.2+cu110
tqdm==4.63.0
trimesh==3.12.7
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.9
uvicorn==0.20.0
wandb==0.12.11
websockets==10.4
Werkzeug==2.0.3
wrapt==1.12.1
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
